refactor: move ApiServiceCockpit in cockpit.services package
