+++ gravitee-apim-rest-api/gravitee-apim-rest-api-service/src/main/java/io/gravitee/rest/api/service/cockpit/services/ApiServiceCockpit.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.gravitee.rest.api.service;
+package io.gravitee.rest.api.service.cockpit.services;
 
 import io.gravitee.rest.api.model.api.ApiEntity;
 import io.gravitee.rest.api.service.cockpit.model.DeploymentMode;
